Почти релиз! Много чего пофикшено, меню полностью доделано.
