+++ data/map1.csv
@@ -1,11 +1,11 @@
-﻿0;4;4;4;4;11_11;4;4;4;4;7
-0;5;7;0;0;0_2;0;2;2;0;0
-1;2;2;11_1;11_1;11_1;11_1;11_1;11;11_1;11_1
+﻿0;4;4;4;4;9;4;4;4;4;7
+0;5;0_2;0;0;0_2;0;2;2;0;0
+1;2;2;11_1;11_1;11_1;11_1;11.1;11.2;11.3;11_1
 1;3;3;0;0_2;0;0_2;0;2;2;0
 0;0_2;0;0;0;3;0;0;1;0;0
 3;0;4;4;4;3;4;4;4;0;3
 0;0;1;0;0;3;0;0;0;0_2;0
 0;2;2;0;0;0_2;0;0;3;3;1
-11_1;11_1;12;11_1;11_1;11_1;11_1;11_1;2;2;1
-0_2;0;2;2;0;0;0_2;0;8;6;0
-8;4;4;4;4;11_11;4;4;4;4;0_2
+11_1;12.1;12.2;12.3;11_1;11_1;11_1;11_1;2;2;1
+0_2;0;2;2;0;0;0_2;0;0;6;0
+8;4;4;4;4;10;4;4;4;4;0_2
